HotFix Valores Decimais Valores

+++ src/Veiculo.java
@@ -141,7 +141,7 @@ public class Veiculo {
         return valor;
     }
 
-    public void setValor(int valor) {
+    public void setValor(Double valor) {
         this.valor = valor;
     }
